Artichoke V1.03
Added ability to have multiple projects
Added ability to change slides on the animation screen
Changed delay in gifs
Removed screenshotting with space and creating gifs with s, replaced it
with the command line.
Current command line commands are:
 /addSlide: Adds a new slide to the end of the project.
/save: Turns all slides into pngs, serializes the project, and creates
a .gif of the project.

A bunch of various other changes that I forgot.